requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
DateTime==4.3
future==0.17.1
idna==2.8
psycopg2-binary==2.8.3
pycparser==2.19
python-telegram-bot==11.1.0
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3
zope.interface==4.6.0
Telethon==1.9.0

